Rename the 'web' entrypoint to 'dom'.
