+++ packages/webcomponentsjs/src/entrypoints/webcomponents-sd-ce-pf-index.js
@@ -15,7 +15,7 @@ subject to an additional IP rights grant found at http://polymer.github.io/PATEN
  */
 
 import './webcomponents-js-index.js';
-import './webcomponents-web-index.js';
+import './webcomponents-dom-index.js';
 import '../unresolved.js';
 import '../flag-parser.js';
 import '../../node_modules/@webcomponents/template/template.js';
